+++ Client/include/CheckBox.hh
@@ -5,7 +5,7 @@
 // Login   <brigno@epitech.net>
 //
 // Started on  Tue Nov 19 13:02:42 2013 brigno
-// Last update Fri Nov 22 00:09:44 2013 brigno
+// Last update Fri Nov 22 05:05:33 2013 brigno
 //
 
 #ifndef		__CHECKBOX_HH__
@@ -17,12 +17,12 @@ class		CheckBox : public AWidget
 {
   int		_states;
   sf::Sprite	_image;
-  sf::Vector2i	_pos;
+  sf::Vector2f	_pos;
   std::string	_name;
   int		_nb;
 
 public:
-  CheckBox(const sf::Event &, const std::string &name, const sf::Vector2i &posTopLeft, const sf::Vector2i &focusTopLeft, const sf::Vector2i &focusBotRight, int nb, bool defaut);
+  CheckBox(const sf::Event &, const std::string &name, const sf::Vector2f &posTopLeft, const sf::Vector2f &focusTopLeft, const sf::Vector2f &focusBotRight, int nb, bool defaut);
   ~CheckBox();
 
 private:
@@ -33,7 +33,7 @@ private:
 public:
   const int		&getStates(void) const;
   const sf::Sprite	&getSprite(void) const;
-  const sf::Vector2i	&getPos(void) const;
+  const sf::Vector2f	&getPos(void) const;
   const std::string	&getName(void) const;
   const int		&getNb(void) const;
   MenuWindow::Status	onFocus(void);
