+++ Client/include/Text.hh
@@ -5,7 +5,7 @@
 // Login   <brigno@epitech.net>
 //
 // Started on  Thu Oct 24 10:55:11 2013 brigno
-// Last update Thu Nov 21 20:45:58 2013 brigno
+// Last update Fri Nov 22 05:05:32 2013 brigno
 //
 
 #ifndef		__TEXT_HH__
@@ -31,10 +31,10 @@ private:
   int		_flag;
 
 public:
-  Text(const std::string &fontPath, const std::string &name, const sf::Event &, const sf::Vector2i &posTopLeft,
-       const sf::Vector2i &focusTopLeft, const sf::Vector2i &focusBotRight, const size_t &size, const bool &enable);
-  Text(const std::string &fontPath, const std::string &name, const sf::Event &, const sf::Vector2i &posTopLeft,
-       const sf::Vector2i &focusTopLeft, const sf::Vector2i &focusBotRight, const size_t &size, const bool &enable, const std::string &);
+  Text(const std::string &fontPath, const std::string &name, const sf::Event &, const sf::Vector2f &posTopLeft,
+       const sf::Vector2f &focusTopLeft, const sf::Vector2f &focusBotRight, const size_t &size, const bool &enable);
+  Text(const std::string &fontPath, const std::string &name, const sf::Event &, const sf::Vector2f &posTopLeft,
+       const sf::Vector2f &focusTopLeft, const sf::Vector2f &focusBotRight, const size_t &size, const bool &enable, const std::string &);
   Text	&operator=(const Text &other);
 
 private:
