+++ Client/sources/CheckBox.cpp
@@ -5,7 +5,7 @@
 // Login   <brigno@epitech.net>
 //
 // Started on  Thu Oct 31 16:22:29 2013 brigno
-// Last update Fri Nov 22 04:51:57 2013 brigno
+// Last update Fri Nov 22 05:05:35 2013 brigno
 //
 
 #include	"CheckBox.hh"
@@ -13,7 +13,7 @@
 #include	"TextureManager.hh"
 #include	"Texture.hh"
 
-CheckBox::CheckBox(const sf::Event &ev, const std::string &name, const sf::Vector2i &posTopLeft, const sf::Vector2i &focusTopLeft, const sf::Vector2i &focusBotRight, int nb, bool defaut) :
+CheckBox::CheckBox(const sf::Event &ev, const std::string &name, const sf::Vector2f &posTopLeft, const sf::Vector2f &focusTopLeft, const sf::Vector2f &focusBotRight, int nb, bool defaut) :
   AWidget(ev, name, posTopLeft, focusTopLeft, focusBotRight, AWidget::CHECKBOX)
 {
   this->_name = name;
@@ -39,7 +39,7 @@ const int		&CheckBox::getNb() const
   return (this->_nb);
 }
 
-const sf::Vector2i	&CheckBox::getPos() const
+const sf::Vector2f	&CheckBox::getPos() const
 {
   return (this->_pos);
 }
